requirements.txt
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
asgiref==3.8.1
base==0.0.0
certifi==2025.1.31
channels==4.2.2
charset-normalizer==3.4.1
colorama==0.4.6
distro==1.9.0
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
mysqlclient==2.2.7
numpy==2.2.4
openai==1.70.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pydantic==2.11.1
pydantic_core==2.33.0
PyJWT==2.10.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
pytz==2025.2
requests==2.32.3
sendgrid==6.11.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
tesseract==0.1.3
tika==3.1.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.1
Unidecode==1.3.8
urllib3==2.3.0
